requirements.txt
Django==4.2.28
djangorestframework==3.14.0
pymysql==1.1.0
python-dotenv
django-cors-headers
mysqlclient
